bad problem solved

+++ src/main/java/com/robosh/web/command/account/ShowAllDriverOrdersCommand.java
@@ -1,6 +1,7 @@
 package com.robosh.web.command.account;
 
-import com.robosh.utils.LoginedUserUtils;
+
+import com.robosh.myUtils.LoginedUserUtils;
 import com.robosh.web.command.Command;
 import com.robosh.web.command.RoutesJSP;
 import com.robosh.model.entity.Order;
@@ -22,7 +23,7 @@ public class ShowAllDriverOrdersCommand implements Command {
     @Override
     public String execute(HttpServletRequest request, HttpServletResponse response) {
         HttpSession session = request.getSession();
-        int idDriver = (int) LoginedUserUtils.getLoginedUser(session).getPersonId();
+        int idDriver =  LoginedUserUtils.getLoginedUser(session).getPersonId();
         int pageNumber = 0;
         int totalNumberRecords = (int) orderService.getAllOrdersCount(idDriver);
         int recordPerPage = 5;
